requirements.txt
uvicorn
fastapi
pyyaml
aiosmtplib
python-dotenv
